refactor(coordinator): 抽出 ime_insertion 子模块
机械式行为保持拆分：从 coordinator.rs 搬出 Windows IME 会话/TSF 插入/非 TSF 回退，
以及焦点目标抓取恢复、前台 App 抓取、capsule 窗口 no-activate 显示/隐藏等 OS 级窗口
插入管线。大量 unsafe + Win32/objc2 FFI 与 cfg 分支、SAFETY/issue 注释原样跟随。
PreparedWindowsImeSessionSlot 留在父模块（Inner 字段类型）。
